Mover eventos de lugar funcionando
>router.js ahora recibe "/events/update", y además app.js lo envía al
ser llamado eventDrop (la función no estaba en el código recibido de
NextU)
>RegistroId cambiado por UltimaIdEvento, este modelo registra la ultima
id utilizada según el usuario, en lugar de utiliar la misma para todos
>Varios get y post en router.js ahora verifican si la sesión del usuario
está abierta, y envían "loginExpira" como respuesta en caso contrario.
app.js recibe los "loginExpira" si se da el caso y redirige al login
(index.html)

+++ Agenda_NodeJS/server/router.js
@@ -60,6 +60,10 @@ router.get("/events/all", function(req, res){
 });
 
 router.post("/events/delete/:id", function(req, res){
+  if(!req.session.logueado){
+    res.send("loginExpira");
+    return;
+  }
   Evento.deleteOne({id:req.params.id}).exec(function(error){
     if(error){
       res.json(error);
@@ -75,7 +79,7 @@ router.post("/events/new", function(req, res){
     return;
   }
 
-  ManagerIds.getIdNuevoEvento(function(idEvento){
+  ManagerIds.getIdNuevoEvento(req.session.usuario, function(idEvento){
     let eventoNuevo = new Evento({
       id:idEvento,
       usuario:req.session.usuario,
@@ -96,6 +100,31 @@ router.post("/events/new", function(req, res){
         });
       }
     });
+  });
+});
+
+router.post("/events/update", function(req, res){
+  if(!req.session.logueado){
+    res.send("loginExpira");
+    return;
+  }
+
+  let eventoActualizado = {
+    usuario:req.session.usuario,
+    titulo:req.body.title,
+    inicio:req.body.start,
+    fin:req.body.end
+  };
+  Evento.updateOne({id:req.body.id}, eventoActualizado, function(error, evento){
+    if(error){
+      res.status(500);
+      res.json(error);
+    } else {
+      res.json({
+        msg:"Evento "+eventoActualizado.titulo+" modificado exitosamente.",
+        idEvento:req.body.id
+      });
+    }
   });
 });
